feat: API for purchasing products
Simple API for purchasing products. It makes a request to the product-service to update the stock for the requested products.

Still to be done will be requesting the product-service to check stock value before updating the stock value, and also the next steps to generate an actual order and have the API return that information instead of the current response.