feat: Add database for storing product reservations
A database and table for storing product reservations. This is to be used by the service to calculate whether a product is in stock to be purchased. It will store the quantity that is "reserved" for a particular product, so if two purchase requests come in sequentially, a product can not be over sold