refactor: Improve error handling
Propagate the errors to the controller advice, to ensure it returns the correct response and message. Previously some errors were wrapped in the wrong exception e.g. they were all wrapped as ValidationExceptions which would return a 400, when it should be a 500.

Also resolved a bug where the reserved products weren't getting cleared from the database when the request failed due to a validation error. They will now be cleared regardless of whether the validations pass or fail